Extract Human Activity Recognition Pipeline Inside a Seprate File
